chore(web): splite file
