+++ src/web.ts
@@ -1,5 +1,5 @@
 import { WebPlugin } from '@capacitor/core';
-import type { ConfirmCardPaymentData, Stripe } from '@stripe/stripe-js';
+import type { ConfirmCardPaymentData, Stripe ,PaymentMethod} from '@stripe/stripe-js';
 
 import type {
   AccountParams,
@@ -32,101 +32,8 @@ import type {
   AvailabilityResponse,
 } from './definitions';
 import { CardBrand } from './definitions';
+import { formBody, _stripePost, _stripeGet } from './helper';
 
-function flatten(json: any, prefix?: string, omit?: string[]): any {
-  let obj: any = {};
-
-  for (const prop of Object.keys(json)) {
-    if (
-      typeof json[prop] !== 'undefined' &&
-      json[prop] !== null &&
-      (!Array.isArray(omit) || !omit.includes(prop))
-    ) {
-      if (typeof json[prop] === 'object') {
-        obj = {
-          ...obj,
-          ...flatten(json[prop], prefix ? `${prefix}[${prop}]` : prop),
-        };
-      } else {
-        const key = prefix ? `${prefix}[${prop}]` : prop;
-        obj[key] = json[prop];
-      }
-    }
-  }
-
-  return obj;
-}
-
-function stringify(json: any): string {
-  let str = '';
-  json = flatten(json);
-
-  for (const prop of Object.keys(json)) {
-    const key = encodeURIComponent(prop);
-    const val = encodeURIComponent(json[prop]);
-    str += `${key}=${val}&`;
-  }
-
-  return str;
-}
-
-function formBody(json: any, prefix?: string, omit?: string[]): string {
-  json = flatten(json, prefix, omit);
-  return stringify(json);
-}
-
-async function _callStripeAPI(fetchUrl: string, fetchOpts: RequestInit) {
-  const res = await fetch(fetchUrl, fetchOpts);
-
-  let parsed;
-
-  try {
-    parsed = await res.json();
-  } catch (e) {
-    parsed = await res.text();
-  }
-
-  if (res.ok) {
-    return parsed;
-  } else {
-    throw parsed?.error?.message ? parsed.error.message : parsed;
-  }
-}
-
-async function _stripePost(
-  path: string,
-  body: string,
-  key: string,
-  extraHeaders?: any,
-) {
-  extraHeaders = extraHeaders || {};
-
-  return _callStripeAPI(`https://api.stripe.com${path}`, {
-    body: body,
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Accept': 'application/json',
-      'Authorization': `Bearer ${key}`,
-      'Stripe-version': '2020-03-02',
-      ...extraHeaders,
-    },
-  });
-}
-
-async function _stripeGet(path: string, key: string, extraHeaders?: any) {
-  extraHeaders = extraHeaders || {};
-
-  return _callStripeAPI(`https://api.stripe.com${path}`, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': `Bearer ${key}`,
-      'Stripe-version': '2020-03-02',
-      ...extraHeaders,
-    },
-  });
-}
 
 export class StripeWeb extends WebPlugin implements StripePlugin {
   private publishableKey: string | undefined;
@@ -483,7 +390,11 @@ class CustomerSession {
   }
 
   async listPm(): Promise<CustomerPaymentMethodsResponse> {
-    const res = await _stripeGet(
+    const res = await _stripeGet<{
+      sources: {
+        data: PaymentMethod[]
+      }
+    }>(
       `/v1/customers/${this.customerId}`,
       this.key.secret,
     );
